requirements.txt
discord.py==2.1.0
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg-pool==3.1.6
aiohttp==3.8.3